requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
dnspython==2.8.0
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
requests==2.33.1
rich==14.3.3
urllib3==2.6.3
